src/main/java/divide_conquer/search_sort/IntroSort.java
package divide_conquer.search_sort;

import java.util.Random;

/**
 * Combination of Heap-, Quick and Insertionsort.
 * - Heap: when getting into deep recursions to prevent QuickSort worst-case
 * - Insert: when the amount of elements is small
 *
 * Max. Recursion depth: 2*logn(N) (if reached -> switch to heapsort)
 * If partition size to small Quick-Sort decays to Insertion-sort (cutoff at 16)
 *
 * The quick-sort additionally uses the media-of-three strategy and choses the first pivot randomly
 *
 *
 * @author Maksim Sandybekov
 * @date 2019-12-16
 */
public class IntroSort<T extends Comparable> {


    private int depthLimit;
    private int partitionSize; // For switching to Insertion-sort
    private int nextPivotIndx; //
    private Random randomizer;


    public IntroSort() {
        this.partitionSize = 16;
    }

    // ----------------------------------
    // Sort-Integer-Arrays
    // ----------------------------------


    public int[] sort(int[] items) {

        if (items.length <= 1) {
            return items;
        }

        int depthLimit = (int) (2 * Math.log(items.length)); // Depth limit to switch to heap-sort
        this.recurseIntegerSort(items, 0, items.length - 1, depthLimit);
        return items;
    }


    /**
     * Divide and Conquer step of Intro-Sort.
     *
     * @param items - items to sort.
     * @param left - left border.
     * @param right - right border
     */
    private void recurseIntegerSort(int[] items, int left, int right, int depthLimit) {

        if ((right - left) <= this.getPartitionSize()) {
            // Change to Insertion-Sort (less than 16-items in sub-problem)
        }

        // Depth Limit reached, switch to HeapSort
        if (depthLimit == 0) {

        }


        if (left > right) {
            return;
        }

    }


    /**
     * Sort a sub-array into left & right parts
     *
     * @param items - items to sort
     * @param left - left border from which to start the sorting
     * @param right - right border marks the point to which all elements should be partitioned
     */
    private void partition(int[] items, int left, int right) {

        int pivotIndx = this.getNextPivotIndx();
        if (pivotIndx == -1) {
            Random rand = this.getRandomizer();
            pivotIndx = rand.nextInt(right-left) + left;
        }

        // Partition item values
        for (int i = left; i < right; i++) {



        }
    }


    /**
     * Sort a sub-array in-place with insertion-sort.
     *
     * @param items - items to sort
     * @param left - left border of items to sort
     * @param right - right border of items to sort
     */
    private void insertionSort(int[] items, int left, int right) {


        for (int i = left; i < items.length; i++) {

            // Sort items till this position
            for (int j = left; j >= 0; j--) {

            }
        }
    }


    /**
     * Shift items
     *
     * @param items
     * @param afterIndx
     * @param tillIndx
     */
    private void shiftAfter(int[] items,  int afterIndx, int tillIndx) {

    }


    /**
     * Swap operations of Intro-Sort for integer values.
     *
     * @param items - items to sort
     * @param from - swap item from index.
     * @param to - swap item to this index.
     */
    private void swap(int[] items, int from , int to) {
        int temp = items[from];
        items[from] = items[to];
        items[to] = temp;
    }


    // ---------------------------------
    // Sort-Generic items
    // ---------------------------------

    /**
     * Sorts a set of items and returns the sorted set.
     *
     * @param items - set of unsorted items
     * @return sorted items
     */
    public T[] sort(T[] items) {

        if (items.length <= 1) {
            return items;
        }


        this.recurseSortItems(items, 0, items.length - 1);
        return items;
    }


    private void recurseSortItems(T[] items, int left, int right) {

        // Small enough partition size to perform Insertion-Sort
        if (this.getPartitionSize() > (right - left)) {
            this.insertionSort(items, left, right);
            return;
        }
    }


    /**
     * Insertion sort.
     *
     * @param items
     * @param left
     * @param right
     */
    private void insertionSort(T[] items, int left, int right) {

        //
        int pivotIndex = this.getNextPivotIndx();
        if (pivotIndex == -1) {
            Random randomizer = this.getRandomizer();
            pivotIndex = randomizer.nextInt(right - left) + left;
        }

        for (int i = left; i < right; i++) {
            break;
        }
    }


    /**
     * Perform an heap-sort on a set of items.
     *
     * @param items - items to be sorted
     */
    private void heapSort(T[] items) {

    }



    private void swapItems(T[] items, int from, int to) {
        T tmp = items[to];
        items[to] = items[from];
        items[from] = tmp;
    }


    // -----------------------
    // Setter/-Getter
    // -----------------------

    private void setPartitionSize(int partitionSize) {
        this.partitionSize = partitionSize;
    }

    private void setDepthLimit(int depthLimit) {
        this.depthLimit = depthLimit;
    }

    private void setNextPivotIndx(int nextPivotIndx) {
        this.nextPivotIndx = nextPivotIndx;
    }

    private int getDepthLimit() {
        return this.depthLimit;
    }

    private int getPartitionSize() {
        return this.partitionSize;
    }

    private int getNextPivotIndx() {
        return this.nextPivotIndx;
    }

    private Random getRandomizer() {
        return this.randomizer;
    }
}
